backend/pyproject.toml
[project]
name = "tanay-backend"
version = "0.1.0"
description = "Backend API for Tzelem multi-agent orchestration platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "slowapi>=0.1.9",
    "pipecat-ai>=0.0.80",
    "pipecat-ai[daily]",
    "aiohttp==3.12.15",
    "python-dotenv==1.0.1",
    "pydantic>=2.11.5",
    "pydantic-settings==2.7.1",
    "onnxruntime>=1.22.1",
    "pipecat-ai-flows>=0.0.18",
    "convex==0.6.0",
    "agentmail>=0.0.53",
    "requests>=2.31.0",
    "langchain>=0.3.22",
    "browserbase",
    "browser-use>=0.5.9",
    "langchain-openai>=0.3.29",
    "browser-use-sdk>=1.0.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "DJ", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["E501", "S101", "PLR0913", "PLR0912", "PLR0911", "PLR0915", "C901", "PLR2004", "S104", "S105", "S106", "G004", "TRY401", "FBT003", "E402", "N806", "SIM117", "N815", "TRY301", "COM812", "T201", "BLE001", "TRY400", "B904", "W293", "PLC0415", "TRY300", "RUF013", "N999", "ARG001", "ARG002", "DTZ005", "ERA001"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["pipecat.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["test_mail_send", "test_mail_comprehensive", "test_mail_batch", "test_mail_performance", "tests.*"]
ignore_errors = true
